fastapi_backend/requirements/base.txt
# FastAPI
fastapi>=0.100.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
starlette>=0.27.0
uvicorn>=0.21.1,<0.22.0
email-validator>=2.0.0,<3.0.0

# Database
sqlalchemy[asyncio]>=2.0.0,<3.0.0
alembic>=1.10.0,<1.11.0
asyncpg>=0.27.0,<0.28.0
psycopg2-binary>=2.9.5,<2.10.0

# Security
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
bcrypt>=4.0.1,<4.1.0
python-multipart>=0.0.6,<0.0.7

# Background tasks
celery>=5.2.7,<5.3.0
redis>=4.5.4,<4.6.0

# Logging
loguru>=0.7.0,<0.8.0

# Template engine
jinja2>=3.1.2,<3.2.0

# Utils
python-dotenv>=1.0.0,<1.1.0
httpx>=0.24.0,<0.25.0
tenacity>=8.2.2,<8.3.0 

# alembic
python-jose[cryptography]>=3.3.0,<3.4.0

# fastapi-users
fastapi-users-db-sqlalchemy
